clean-up SeaHasher.Combiner code style

+++ Haschisch/Hashers/SeaHasher.cs
@@ -201,19 +201,19 @@ namespace Haschisch.Hashers
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             public int Combine<T1>(T1 value1)
             {
-                var v1 = value1?.GetHashCode() ?? 0;
+                var v1 = (uint)(value1?.GetHashCode() ?? 0);
                 SeaHashSteps.Initialize(out var a, out var b, out var c, out var d);
-                SeaHashSteps.MixFinalPartialStep(ref a, (uint)v1);
+                SeaHashSteps.MixFinalPartialStep(ref a, v1);
                 return (int)SeaHashSteps.Finish(ref a, ref b, ref c, ref d, sizeof(int));
             }
 
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             public int Combine<T1, T2>(T1 value1, T2 value2)
             {
-                var v1 = value1?.GetHashCode() ?? 0;
-                var v2 = value2?.GetHashCode() ?? 0;
+                var v1 = (uint)(value1?.GetHashCode() ?? 0);
+                var v2 = (uint)(value2?.GetHashCode() ?? 0);
                 SeaHashSteps.Initialize(out var a, out var b, out var c, out var d);
-                SeaHashSteps.MixFinalPartialStep(ref a, ((uint)v2 << 32) | (uint)v1);
+                SeaHashSteps.MixFinalPartialStep(ref a, ((ulong)v2 << 32) | v1);
                 return (int)SeaHashSteps.Finish(ref a, ref b, ref c, ref d, 2 * sizeof(int));
             }
 
@@ -235,35 +235,35 @@ namespace Haschisch.Hashers
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             public int Combine<T1, T2, T3, T4>(T1 value1, T2 value2, T3 value3, T4 value4)
             {
-                var v1 = value1?.GetHashCode() ?? 0;
-                var v2 = value2?.GetHashCode() ?? 0;
-                var v3 = value3?.GetHashCode() ?? 0;
-                var v4 = value4?.GetHashCode() ?? 0;
+                var v1 = (uint)(value1?.GetHashCode() ?? 0);
+                var v2 = (uint)(value2?.GetHashCode() ?? 0);
+                var v3 = (uint)(value3?.GetHashCode() ?? 0);
+                var v4 = (uint)(value4?.GetHashCode() ?? 0);
                 SeaHashSteps.Initialize(out var a, out var b, out var c, out var d);
                 SeaHashSteps.MixFinalPartialStep(
                     ref a,
                     ref b,
-                    (uint)v1 | ((ulong)v2 << 32),
-                    (uint)v3 | ((ulong)v4 << 32));
+                    v1 | ((ulong)v2 << 32),
+                    v3 | ((ulong)v4 << 32));
                 return (int)SeaHashSteps.Finish(ref a, ref b, ref c, ref d, 4 * sizeof(int));
             }
 
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             public int Combine<T1, T2, T3, T4, T5>(T1 value1, T2 value2, T3 value3, T4 value4, T5 value5)
             {
-                var v1 = value1?.GetHashCode() ?? 0;
-                var v2 = value2?.GetHashCode() ?? 0;
-                var v3 = value3?.GetHashCode() ?? 0;
-                var v4 = value4?.GetHashCode() ?? 0;
-                var v5 = value5?.GetHashCode() ?? 0;
+                var v1 = (uint)(value1?.GetHashCode() ?? 0);
+                var v2 = (uint)(value2?.GetHashCode() ?? 0);
+                var v3 = (uint)(value3?.GetHashCode() ?? 0);
+                var v4 = (uint)(value4?.GetHashCode() ?? 0);
+                var v5 = (uint)(value5?.GetHashCode() ?? 0);
                 SeaHashSteps.Initialize(out var a, out var b, out var c, out var d);
                 SeaHashSteps.MixFinalPartialStep(
                     ref a,
                     ref b,
                     ref c,
-                    (uint)v1 | ((ulong)v2 << 32),
-                    (uint)v3 | ((ulong)v4 << 32),
-                    (uint)v5);
+                    v1 | ((ulong)v2 << 32),
+                    v3 | ((ulong)v4 << 32),
+                    v5);
                 return (int)SeaHashSteps.Finish(ref a, ref b, ref c, ref d, 5 * sizeof(int));
             }
 
@@ -271,20 +271,20 @@ namespace Haschisch.Hashers
             public int Combine<T1, T2, T3, T4, T5, T6>(
                 T1 value1, T2 value2, T3 value3, T4 value4, T5 value5, T6 value6)
             {
-                var v1 = value1?.GetHashCode() ?? 0;
-                var v2 = value2?.GetHashCode() ?? 0;
-                var v3 = value3?.GetHashCode() ?? 0;
-                var v4 = value4?.GetHashCode() ?? 0;
-                var v5 = value5?.GetHashCode() ?? 0;
-                var v6 = value6?.GetHashCode() ?? 0;
+                var v1 = (uint)(value1?.GetHashCode() ?? 0);
+                var v2 = (uint)(value2?.GetHashCode() ?? 0);
+                var v3 = (uint)(value3?.GetHashCode() ?? 0);
+                var v4 = (uint)(value4?.GetHashCode() ?? 0);
+                var v5 = (uint)(value5?.GetHashCode() ?? 0);
+                var v6 = (uint)(value6?.GetHashCode() ?? 0);
                 SeaHashSteps.Initialize(out var a, out var b, out var c, out var d);
                 SeaHashSteps.MixFinalPartialStep(
                     ref a,
                     ref b,
                     ref c,
-                    (uint)v1 | ((ulong)v2 << 32),
-                    (uint)v3 | ((ulong)v4 << 32),
-                    (uint)v5 | ((ulong)v6 << 32));
+                    v1 | ((ulong)v2 << 32),
+                    v3 | ((ulong)v4 << 32),
+                    v5 | ((ulong)v6 << 32));
                 return (int)SeaHashSteps.Finish(ref a, ref b, ref c, ref d, 6 * sizeof(int));
             }
 
@@ -292,23 +292,23 @@ namespace Haschisch.Hashers
             public int Combine<T1, T2, T3, T4, T5, T6, T7>(
                 T1 value1, T2 value2, T3 value3, T4 value4, T5 value5, T6 value6, T7 value7)
             {
-                var v1 = value1?.GetHashCode() ?? 0;
-                var v2 = value2?.GetHashCode() ?? 0;
-                var v3 = value3?.GetHashCode() ?? 0;
-                var v4 = value4?.GetHashCode() ?? 0;
-                var v5 = value5?.GetHashCode() ?? 0;
-                var v6 = value6?.GetHashCode() ?? 0;
-                var v7 = value7?.GetHashCode() ?? 0;
+                var v1 = (uint)(value1?.GetHashCode() ?? 0);
+                var v2 = (uint)(value2?.GetHashCode() ?? 0);
+                var v3 = (uint)(value3?.GetHashCode() ?? 0);
+                var v4 = (uint)(value4?.GetHashCode() ?? 0);
+                var v5 = (uint)(value5?.GetHashCode() ?? 0);
+                var v6 = (uint)(value6?.GetHashCode() ?? 0);
+                var v7 = (uint)(value7?.GetHashCode() ?? 0);
                 SeaHashSteps.Initialize(out var a, out var b, out var c, out var d);
                 SeaHashSteps.MixStep(
                     ref a,
                     ref b,
                     ref c,
                     ref d,
-                    (uint)v1 | ((ulong)v2 << 32),
-                    (uint)v3 | ((ulong)v4 << 32),
-                    (uint)v5 | ((ulong)v6 << 32),
-                    (uint)v7);
+                    v1 | ((ulong)v2 << 32),
+                    v3 | ((ulong)v4 << 32),
+                    v5 | ((ulong)v6 << 32),
+                    v7);
                 return (int)SeaHashSteps.Finish(ref a, ref b, ref c, ref d, 7 * sizeof(int));
             }
 
@@ -316,24 +316,24 @@ namespace Haschisch.Hashers
             public int Combine<T1, T2, T3, T4, T5, T6, T7, T8>(
                 T1 value1, T2 value2, T3 value3, T4 value4, T5 value5, T6 value6, T7 value7, T8 value8)
             {
-                var v1 = value1?.GetHashCode() ?? 0;
-                var v2 = value2?.GetHashCode() ?? 0;
-                var v3 = value3?.GetHashCode() ?? 0;
-                var v4 = value4?.GetHashCode() ?? 0;
-                var v5 = value5?.GetHashCode() ?? 0;
-                var v6 = value6?.GetHashCode() ?? 0;
-                var v7 = value7?.GetHashCode() ?? 0;
-                var v8 = value8?.GetHashCode() ?? 0;
+                var v1 = (uint)(value1?.GetHashCode() ?? 0);
+                var v2 = (uint)(value2?.GetHashCode() ?? 0);
+                var v3 = (uint)(value3?.GetHashCode() ?? 0);
+                var v4 = (uint)(value4?.GetHashCode() ?? 0);
+                var v5 = (uint)(value5?.GetHashCode() ?? 0);
+                var v6 = (uint)(value6?.GetHashCode() ?? 0);
+                var v7 = (uint)(value7?.GetHashCode() ?? 0);
+                var v8 = (uint)(value8?.GetHashCode() ?? 0);
                 SeaHashSteps.Initialize(out var a, out var b, out var c, out var d);
                 SeaHashSteps.MixStep(
                     ref a,
                     ref b,
                     ref c,
                     ref d,
-                    (uint)v1 | ((ulong)v2 << 32),
-                    (uint)v3 | ((ulong)v4 << 32),
-                    (uint)v5 | ((ulong)v6 << 32),
-                    (uint)v7 | ((ulong)v8 << 32));
+                    v1 | ((ulong)v2 << 32),
+                    v3 | ((ulong)v4 << 32),
+                    v5 | ((ulong)v6 << 32),
+                    v7 | ((ulong)v8 << 32));
                 return (int)SeaHashSteps.Finish(ref a, ref b, ref c, ref d, 8 * sizeof(int));
             }
         }
